back-end/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
SQLAlchemy==2.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
